refactor(pdf): colocate metadata parser
