--- pdf-metadata.js
@@ -1,163 +1,0 @@
-const DC_NS = 'http://purl.org/dc/elements/1.1/'
-const RDF_NS = 'http://www.w3.org/1999/02/22-rdf-syntax-ns#'
-const CALIBRE_NS = 'http://calibre-ebook.com/xmp-namespace'
-const CALIBRE_SI_NS = 'http://calibre-ebook.com/xmp-namespace-series-index'
-
-/** @typedef {string | number[] | Uint8Array | null | undefined} PDFMetadataValue */
-/** @typedef {Record<string, PDFMetadataValue>} PDFInfo */
-
-// PDF 32000-1, Table D.2. This mirrors pdf.js's stringToPDFString so Info
-// dictionary values read natively have the same Unicode representation as
-// values returned by PDFDocumentProxy.getMetadata().
-const PDF_STRING_TRANSLATE_TABLE = [
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0x2d8, 0x2c7, 0x2c6, 0x2d9, 0x2dd, 0x2db, 0x2da, 0x2dc,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-    0x2022, 0x2020, 0x2021, 0x2026, 0x2014, 0x2013, 0x192, 0x2044,
-    0x2039, 0x203a, 0x2212, 0x2030, 0x201e, 0x201c, 0x201d, 0x2018,
-    0x2019, 0x201a, 0x2122, 0xfb01, 0xfb02, 0x141, 0x152, 0x160,
-    0x178, 0x17d, 0x131, 0x142, 0x153, 0x161, 0x17e, 0, 0x20ac,
-]
-
-/** @param {number[] | Uint8Array} value */
-const toUint8Array = value => value instanceof Uint8Array ? value : new Uint8Array(value)
-
-const removeEscapeSequences = value => {
-    let result = ''
-    let escaping = false
-    for (const char of value) {
-        if (char.charCodeAt(0) === 0x1b) {
-            escaping = !escaping
-        } else if (!escaping) {
-            result += char
-        }
-    }
-    return result
-}
-
-/** @param {PDFMetadataValue} value */
-export const decodePDFString = value => {
-    if (value == null || typeof value === 'string') return value
-    let bytes = toUint8Array(value)
-    let encoding
-    if (bytes[0] === 0xfe && bytes[1] === 0xff) encoding = 'utf-16be'
-    else if (bytes[0] === 0xff && bytes[1] === 0xfe) encoding = 'utf-16le'
-    else if (bytes[0] === 0xef && bytes[1] === 0xbb && bytes[2] === 0xbf) encoding = 'utf-8'
-    if (encoding) {
-        if (encoding.startsWith('utf-16') && bytes.length % 2 === 1) bytes = bytes.slice(0, -1)
-        try {
-            return removeEscapeSequences(new TextDecoder(encoding, { fatal: true }).decode(bytes))
-        } catch {
-            // Match pdf.js: fall through to PDFDocEncoding when BOM decoding fails.
-        }
-    }
-    const chars = []
-    for (let i = 0; i < bytes.length; i++) {
-        const byte = bytes[i]
-        if (byte === 0x1b) {
-            while (++i < bytes.length && bytes[i] !== 0x1b) {
-                // Skip the escape sequence.
-            }
-            continue
-        }
-        const code = PDF_STRING_TRANSLATE_TABLE[byte]
-        chars.push(String.fromCharCode(code || byte))
-    }
-    return chars.join('')
-}
-
-/** @param {PDFMetadataValue} value */
-const decodeXMP = value => {
-    if (!value) return null
-    const raw = typeof value === 'string'
-        ? value
-        : new TextDecoder('utf-8').decode(toUint8Array(value))
-    return raw.replace(/^[^<]+/, '').replaceAll(/>\\376\\377([^<]+)/g, (all, codes) => {
-        const bytes = codes.replaceAll(/\\([0-3])([0-7])([0-7])/g,
-            (code, d1, d2, d3) => String.fromCharCode(d1 * 64 + d2 * 8 + d3))
-            .replaceAll(/&(amp|apos|gt|lt|quot);/g, (str, name) => ({
-                amp: '&', apos: "'", gt: '>', lt: '<', quot: '"',
-            })[name])
-        const chars = ['>']
-        for (let i = 0; i < bytes.length; i += 2) {
-            const code = bytes.charCodeAt(i) * 256 + bytes.charCodeAt(i + 1)
-            chars.push(code >= 32 && code < 127 && code !== 60 && code !== 62 && code !== 38
-                ? String.fromCharCode(code)
-                : `&#x${(0x10000 + code).toString(16).substring(1)};`)
-        }
-        return chars.join('')
-    })
-}
-
-const parseXMP = raw => {
-    if (!raw) return { doc: null, values: new Map() }
-    let doc
-    try {
-        doc = new DOMParser().parseFromString(raw, 'application/xml')
-    } catch {
-        return { doc: null, values: new Map() }
-    }
-    if (doc.getElementsByTagName('parsererror').length) {
-        return { doc: null, values: new Map() }
-    }
-    const values = new Map()
-    for (const description of doc.getElementsByTagNameNS(RDF_NS, 'Description')) {
-        for (const entry of description.children) {
-            if (entry.namespaceURI !== DC_NS) continue
-            const name = `dc:${entry.localName.toLowerCase()}`
-            if (name === 'dc:creator' || name === 'dc:subject') {
-                const container = entry.firstElementChild
-                const items = container
-                    ? Array.from(container.children)
-                        .filter(item => item.namespaceURI === RDF_NS && item.localName === 'li')
-                        .map(item => item.textContent.trim())
-                    : []
-                values.set(name, items)
-            } else {
-                values.set(name, entry.textContent.trim())
-            }
-        }
-    }
-    return { doc, values }
-}
-
-const getInfo = (info, name) =>
-    decodePDFString(info?.[name] ?? info?.[name.toLowerCase()]) ?? undefined
-
-const getCalibreSeries = doc => {
-    const series = doc?.getElementsByTagNameNS(CALIBRE_NS, 'series').item(0)
-    const name = series?.getElementsByTagNameNS(RDF_NS, 'value').item(0)?.textContent?.trim()
-    if (!name) return null
-    const position = series.getElementsByTagNameNS(CALIBRE_SI_NS, 'series_index')
-        .item(0)?.textContent?.trim()
-    return position ? { name, position } : { name }
-}
-
-/**
- * Normalize PDF Info/XMP into the metadata shape used by the full reader.
- * @param {{ info?: PDFInfo, xmp?: PDFMetadataValue }} [input]
- */
-export const parsePDFMetadata = ({ info = {}, xmp = null } = {}) => {
-    const raw = decodeXMP(xmp)
-    const { doc, values } = parseXMP(raw)
-    const metadata = {
-        title: values.get('dc:title') ?? getInfo(info, 'Title'),
-        author: values.get('dc:creator') ?? getInfo(info, 'Author'),
-        contributor: values.get('dc:contributor'),
-        description: values.get('dc:description') ?? getInfo(info, 'Subject'),
-        language: values.get('dc:language'),
-        publisher: values.get('dc:publisher'),
-        subject: values.get('dc:subject'),
-        identifier: values.get('dc:identifier'),
-        source: values.get('dc:source'),
-        rights: values.get('dc:rights'),
-    }
-    const series = getCalibreSeries(doc)
-    if (series) metadata.belongsTo = { series }
-    return metadata
-}
